+++ src/pits_mras/training/__init__.py
@@ -14,6 +14,7 @@ from pits_mras.training.hinf_minmax import (
 from pits_mras.training.irl_trainer import train_irl_critic
 from pits_mras.training.pretrain import pretrain_pitnn
 from pits_mras.training.sac import SACTrainer
+from pits_mras.training.tdmpc import tdmpc_update
 
 __all__ = [
     "pretrain_pitnn",
@@ -23,4 +24,5 @@ __all__ = [
     "hinf_minmax_from_dynamics",
     "hji_residual",
     "SACTrainer",
+    "tdmpc_update",
 ]